Make library a package
