+++ .gitignore
@@ -1,2 +1,3 @@
 __pycache__/
 .idea/
+musical_spoon.egg-info/
