+++ musical_spoon/__init__.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Sep 22 20:46:57 2019
+
+@author: Arctandra
+"""
+
+#Dieses Modul stellt eine Bibliothek von Funktionen dar, die alle wesentlichen Funktionalitäten des
+#Musical_spoon packages umfasst. Um das Programm in vollem Umfang zu nutzen, sollte der import dieses
+#core.py Moduls ausreichen. Die exakte Funktionsweise, Eingabe- und Ausgabeparameter sind bitte der
+#beigefügten Dokumentation zu entnehmen.
+
+#-------------Basic------------------------------------------------------------
+from .vertex import Vertex
+from .edge import Edge
+from .graph import Graph
+
+from .show_graph import show_graph
+from .show_graph import show_two_graphs
+from .show_graph import show_graph_comparable
+
+#-------------Parsing----------------------------------------------------------
+from .parser2 import parse
+from .parser2 import reverse_parser
+from .parser2 import parse_chem
+
+#-------------Random Graph Erzeugung-------------------------------------------
+from .random_graph import random_graph
+from .random_graph import random_chess_graph
+from .random_graph import random_triangular_graph
+from .random_graph import complete_graph
+from .random_graph import cut_edges
+from .random_graph import add_random_nodes
+
+#-------------Bronk-Kerbosch-Algorithmus---------------------------------------
+from .bronk_pivot import find_ankered_mcis
+from .bronk_pivot import find_mcis
+from .bronk_pivot import find_mcis_without_prompt
+from .bronk_pivot import modular_product
+
+#------------Cordella-Algorithmus----------------------------------------------
+from .Cordella_max import Cordella
+from .Cordella_max import find_successors
+from .Cordella_max import find_predecessors
+
+#------------Guide Tree--------------------------------------------------------
+from .guide_tree import guide_tree
+from .guide_tree import create_tree
+from .guide_tree import traverse_tree
+from .guide_tree import traverse_linear
+from .guide_tree import score_similarity
