+++ musical_spoon/main.py
@@ -1,4 +1,4 @@
-from core import *
+from . import *
 import numpy as np
 import timeit
